pyproject.toml
[project]
name = "multi-im-siglip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "torch>=2.7.0",
]
